Product Reviews now use a VoteForm which is more flexible than the previous vote up/down mechanic. The delta is passed rather than an action field, so if an implementation uses something other than up/down voting it'll still work.
I also made it not care if you try and vote multiple times, it'll just update your vote if it already exists. It'll be the job of the template to decide which voting forms to show or hide

update_totals on reviews doesn't get passed a vote anymore, instead it just calculates fresh each time.