added statistics & graph to dashboard page (#56777)
I updated the statistics on the dashboard index page to provide
more updates on orders received in the last 24hrs and included the
generation of a graph to illustrate the hourly orders over this
time period.
The graph is a complete CSS solution that needs some cleaning up
and tweaking by FEDs but provides the basic functionality.