modified notification models and added tests
I updated the notification models to be more generic. The major
properties such as ``User``, ``email`` and ``date_created`` are
stored in an abstract class that allows for implementing other
notifications at a later point in time.
In addition, I made ``email`` and ``user`` optional as only one
of the two should be set in my opinion: ``email`` for anonymous
users and ``user`` for registered users which will provide the
email as part of the profile.

At the current state, the test only implement views related
to registered users.