implemented confirmation and unsubscription incl. tests
I added dates to the model of to be able to track stale notifications
based on creation and modification dates.

I removed the persistence key from the ``AbstractNotification`` model
and removed the middleware as I do not think that it is required. I
discussed this with Jon Moss and we both think that having a middleware
that processes every HTTP request. As the functionality in the
middleware only applied to anonymous users this seems to be overhead.
But please feel free to convince me of the opposite.