added sending of notification emails on stock record update
I added the backend funtionality for sending notification emails
when stock records are updated. For now, sending the messages is
triggered by the ``post_save`` signal of the ``StockRecord``
model which means that the user will receive multiple messages
if subscribed to multiple products that are in the same update,
e.g. a batch import. The reason for that is that implementing
to collect batch updates and send one update message per user
can be quite tricky, I think. This needs a proper underlying
concept that can be added at some point.

The current implementation sends out a notification to each
registered and unregistered user that has a notification in
state ``ACTIVE``. Each notification is marked ``INACTIVE``
and stamped with the current date as ``date_notified`` property
for future reference after sending out the emails.