updated product browsing page to work with notification
I fixed some minor issues with notifications on the product list
page (main product view). The notification form was not passed
to the template and therefore would not be able to add a notification
as the email was missing. This is fixed now.

I also removed the "Add To Basket" button for products that are not
in stock. This is replaced by the "notify me" button.

Some FE optimisation has to take place to make it a bit more shiny
but the functionality is there.