added templatetag for notification signup info
I added a template tag for notifications to determine if a user
has already signed up to receive notifications for a specific
product. The tag provides a boolean flag in the template context
that can be used to show/hide the notification button & form if
a user has already signed up.