added clean up command for unconfirmed notifications
I have added a management command to clean up unconfirmed
notifications that are older then a certain date. The expiry date
is calculated relative to the current date and time and can be
specified in ``days`` or ``hours`` (or both). A valid clean up
command could look like this:

    ./manage.py oscar_cleanup_notifications --days 1 --hours 12

Currently this will remove all notifications directly derived from
``AbstractNotification``. This must be adjusted after a decision
has been made on how to handle the derived classed from abstract
bases (InheritanceManager, django-model-utils, __subclasses__).

I also added a ``tests.py`` to the commands folder to test the
functionality of the command.