added model_utils and made notifications more generic
I added the model_utils django package to Oscar and updated the
implementation of the notifications app to use the
``InheritanceManager``. To make the notifications work with the
more generic approach, I moved the abstract base notification to
``abstract_models.py`` and introduced a non-abstract base class
``Notification`` to be able to us the inheritance manager.
I also added a method to the model that provides the notification
item of a notification, e.g. the product for a
``ProductNotification``. This allows for using a more generic
approach in the templates using the
``generate_notification_description`` tag which renders a template
for each notification type to make it possible to customise the
description of a notification in the template, e.g. add a link to
the product page for a product notification.