fixed problem with select_related on products without stock records
I realise that the ``select_related`` on the product - added to
reduce the amount SQL queries - does cause problems with products
that have no stock record. Especially, for product variants this
is a huge issue.
I have reverted this change as there is no easy way around it. A
possible solution might be to generate the JOIN directly by using
``join`` on the products but this needs some more evaluation.