Updated breadcrumbs on product list pages
The breadcrumbs in the product list page using the ``browse.html``
template only shows the leaf category selected but not the whole
path in the hierachy. This means when selecting a parent
category on a product page, the user will loose track of where
in the tree they actually were.