updated breadcrumbs on product list pages
The breadcrumbs in the product list page using the ``browse.html``
template only shows the leaf category selected but not the whole
path in the hierachy. This means when selecting a parent
category on a product page, the user will loose track of where
in the tree they actually were. Talking to Gareth we updated
the template with the fix included here and thought it might
be good to provide it to Oscar.

This is only a suggestion. Please feel free to close it if you
don't agree with it.