moves the main functionality of notification models into abstract base
The main functionality of the notification model has moved into the
abstract base class adding a class attribute ``item_field_name`` to
allow for subclasses of notifications to provide the item specific
information without re-implementing it.
The ``item_field_name`` specifies the name of the model field used
in the subclass to specify the relation ship to the referenced item.
This field is then used to generate the URLs for unsubscribe and
confirm URLs as well as getting the actual item from the
notification.

The migrations are merged into one initial migration as part of this
commit.