adds a new management command for sending email notifications
A new management command is added that checks all products for
active notifications and sends out notification emails for products
that are available to buy.
To keep the implementation DRY, a new ``utils`` module is added to the
notification app that provides the function used to send out emails
and deactivate notifications that have been processed.

In addition, a new setting is added to Oscar that allows for
disabling the default setup to send notifications on stock record
changes. If this setting is disabled, the only way to send out
notifications is to run the ``oscar_send_notifications`` management
command as a cronjob or something similar.