adding a convenient extension for WebTest
I've added a convenient wrapper around the WebTest class that I
have used in a few projects recently. I found it quite helpful. I
am not sure if this is something that is helpful to have in core
Oscar but thought I'd put it in a pull request for discussion.
Any thoughts?