Finished the report proof of concept
There is still some big questions marks on this implementation related
with the Generator and ReportView objects and how they interact
together.

The ReportView.form_valid and Generator.run currently have some
cross-logic which might is not good in case if a user needs to
overide one of this methods.