Use CSS over Less as the default in the sandbox
You only want to compile Less files if you are developing the CSS of
Oscar.  Most people don't do this and so we default to serving
compressed CSS without Less compilation.  This also means that
the many contributors don't need to bother installing node.js and lessc,
which makes the installation process much simpler.

This change also re-introduces the `USE_LESS` setting which controls
whether Less files are used or not.