Use absolute image URLs in LESS files
Relative URLs (eg '../img/ui/stars.png') are incorrect when the LESS
files are compiled on-the-fly.  When this happens, the compiled CSS file
is located in `/static/CACHE/css` (rather than the usual location of
`/static/oscar/css`).  Hence the relative URL is not correct for images.

This commit simply replaces the "../" with "/static/oscar" which is fine
for the demo site but probably not suitable for core Oscar.  This
isn't a long-term solution as this URL is built with the STATIC_URL
setting which is configurable per project.  Better to have a LESS
variable which defines a "base URL" for included images.