Remove need to have product in context
This fix checks to see if the product is in the context before trying to access its id property. The current logic is that it accesses its id property so that it can remove the current product that is being viewed from the recently viewed product list. If there is no product in the context this check doesn't happen anymore.