Explicitly update category children slugs
When a category's slug is updated, it's children's slugs have to be
updated as well. That already happened due to quirky behavior - both the
move function used in Oscar's tests and the form in the dashboard caused
the children slugs to be updated. But children slugs would still break
when a category was updated directly, e.g. in the shell.
The behavior is now explicit in AbstractCategory's save method. This
probably causes quite a few redundant queries, but shouldn't be a
problem as categories likely don't update often.