Change compressor settings to fix relative URL issue
When the LESS files are served but compression is disabled, the relative
URL between the generated CSS and other assets is wrong (using the
default compressor settings).  This fixes changes the output directory
name to match the path of Oscar's assets, which fixes this issue.

Related to https://github.com/jezdez/django_compressor/issues/226

Fixes #649