AddToBasketForm now handles options properly
When adding to the basket, products with options were treated as
different products for each combination of options. This allowed adding
more products into the basket than are available for purchase.

The handling has been updated to sum up the already added quantity of
the product regardless of options.

Fixes #665