Acceptance tests using Behave, WebTest and Nose work!
The hard work for this is done in django_nose which provides the
appropriate WebTest app, which we can simply re-use in the behave
environment.py.

They need to be run using:

    behave tests/features

Note, you need to use:

    behave --no-capture tests/features

for PDB to work.

First feature is not erroring

Simple homepage tests works

Extra scenario added
# Last 5 commit messages
# ----------------------
# 815b62d Correct formatting error [David Winterbottom]
# 568715b Use compatability layer for user in test [David Winterbottom]
# fcddaec Use bootstrap input styles to shrink quantity field width [David Winterbottom]
# 3b55d29 Fixed the display of children categories in the primary site navigation [Jon Price]
# 661baf5 Set size on wishlist line quantity form field [David Winterbottom]