Make sure all BasketLine instances have a strategy set.
When building the formset for the basket lines the lines are first fetched
directly from the database and without a strategy when django determines
how many forms it has to create. It is these basket line instances that are
used later in the template, where they need a strategy to display price
information.