Add circular import warning to get_classes
Before the fix in e97ec27c958e56055e7383a3ed6a38a49a244512, the new
get_classes threw an AppNotFoundError whereas get_classes prior to the
refactor did not.

The condition is that neither the local module nor the Oscar module
could be imported. Our reasoning is that this is an issue with circular
imports. Before the refactor of get_classes, not being able to import
the Oscar in the shortcut path (beginning of function) would raise an
ImportError; which is then masked by a wrapping get_classes call. Hence,
the circular import is hidden.

This fix highlights the condition explicitly by raising an error that is
intentionally not an ImportError.