The wysiwyg should not be initialized for the raw html body textarea.
Since the wysiwyg is set by default for all dashboard promotions
the raw html body field does not render proper HTML since it gets
processed by the wysiwyg. Need to stop it from initializing.

Added a class to the body field and testing for it in the dashboard js script.
We could also be explicit and add wysiwyg classes to each of the fields that
need to render the wysiwyg intsead of overriding the one.
Could also test for form type in the template before spitting out the 'wysiwyg' class
on the HTML form tag. Not sure how.