Factor out new OrderPreviewView out of PaymentDetailsView
build_submission and get_error_response were the only shared codepaths
and have been moved into the OrderPlacementMixin.
When an error happens in .handle_payment, the logic now issues a
redirect to PaymentDetailsView with kwarg error=True. This is slightly
different from before, but should result in the same template and same
error message.