Slugs: Make forcing ASCII characters optional
Oscar uses unidecode() to turn unicode characters into their ASCII
equivalents where possible. This step was not part of Oscar's default
slugifier function and hence it was not possible to skip it. It now is
part of the default slugifier function, which enables international
URLs.

OSCAR_SLUG_FUNCTION now also accepts a function in string notation.