[Backport] Add helper for Meta.fields definitions for User forms
This is necessary because the Meta def itself is sufficient to crash the
app for a different User model - the form doesn't even need to be
instantiated. There is still plenty more code that makes explicit
reference to first_name and last_name however.

Fixes #1275

Moved to oscar.core.compat and renamed by @maikhoepfel

Conflicts:
	oscar/apps/dashboard/partners/forms.py
	oscar/core/compat.py

Backport of a32766b.