Ensure dashboard is responsive
Before this change, the dashboard wasn't behaving responsively which was
a regression (I think) since 0.6.4. It looks like the change came from
the bootstrap upgrade to 2.3.2.

This change ensures responsive.css is loaded for the dashboard templates
and adjusts the padding on the body so content isn't flush to the
viewport edges.

The link to the dashboard is shown when in the narrowest viewport. It
was previously hidden.

A couple of minor FED gremlins are also squashed (hat-tip @andysellick).