Bump several dependencies for Python 3 compatibility
Debug toolbar 1.2 has some more usable panels, and less trouble with
jQuery interactions. It also is the first version to be compatible with
Django 1.7, which we'll support in the upcoming Oscar release.

django-compressor 1.4 has finally been released and supports Python 3
and comes with lots of other goodies.

sorl-thumbnail seems to be barely maintained anymore. I don't dare
switch to the beta version for Python 2. There's a few critical issues
reported on their issue tracker atm:
https://github.com/mariocesar/sorl-thumbnail/issues?labels=critical&page=1&state=open

django-nose-querycount is not Python 3 compatible and has been moved.