Update strategy mixins concerned with pricing
The NoTax and FixedRateTax have been refactored to show how similar
they are. That also means that the fixes from c3940c3 for zero-price
products are applied in all cases, which wasn't the case!

Both mixins now pass in the tax_rate instead of the tax as per the
updated pricing policy interface.