Run Solr *and* Django within container
The container now runs supervisor to manage a Solr and Django process.
We use the built-in Jetty server for Solr for simplicity and symlink in the
sandbox Solr conf folder.

One problem is that we can't run `rebuild_index` as part of the
Dockerfile provisioning as the Solr process isn't running at that point.
Hence why the 'sandbox' make target has been split in two.
Will have a think about how to solve this.