Improve get_classes matching
When looking at the INSTALLED_APPS entries, Oscar wrongly matched apps
ending with the same string as an Oscar app. The app in question was
named phone2address and came before the custom address app, and classes
weren't picked up as a consequence.

It was reported on the mailing list:
https://groups.google.com/d/msgid/django-oscar/CAJrCdYATiqd317ijwPZ33VmGuWKvzoHZOV16aQJ-J0fBBe_Asg%40mail.gmail.com?utm_medium=email&utm_source=footer