Extract man. cmd. logging behaviour into base class
A bit of a sprawling commit this one. The main change is to extract the
boilerplate logging code into the `OscarBaseCommand` class. This
instantiates the appropriate logger instance and logs the start and
finish of a command.

The alerts module also is extended to allow a logger to be passed to
each function. Some extra logging calls are included too.