Handle app_labels for custom users
The app label needs to be the same that is used in the AUTH_USER_MODEL
setting to specify the custom user. For Oscar's AbstractUser, we
shouldn't set an app label; Django's autodetection should work in most
cases.

It in facts works for the test suite's custom user as well; but as we
specify app labels everywhere, we might as well specify it there.