Filter product update template by product structure
Several sections are not applicable for products of a certain structure.
Hence, they're naively hidden by hiding them in the navigation.
No further hiding is done to ease customisation of what gets shown, to
reduce redundancy and to reduce the chance of the nav and the actual
section display logic getting out of sync.

It's recommended to view this commit with the ignore-whitespace option.