Simplify product list queryset description
The previous implementation was too tightly coupled to the actual search
process, and rather verbose. It also had issues for translators in the
case of searching for both UPC and title.

This commit simplifies the description handling and provides a hook to
plug in more advanced descriptions. But I think a visual indication of
"those results are filtered" is a lot more useful than any description,
no matter how detailed.

Furthermore, this commit swaps out the "Create product" and "Search
products" sections to visually tie the search and search results closer
together.