Workaround for pysolr not supporting Solr 4 error messages
Generally, Haystack now seems to clean the facet queries. But there's a
UnicodeBug lingering that is triggered because pysolr 3.1 doesn't
support Solr 4's JSON error messages. Support for that has just been
added on the master branch, but the potential for the UnicodeDecodeError
is still there until my PR #127 is merged.

Closes #1060.