Avoid excess query
Iterating objects manually here avoids an excess query that developers can't optimize out of the method. This could (**should?**) be broken out to another function, and the error message might not follow Oscar convention so I'm wanting to know what the "right way" to do this in Oscar would be.