Address model loading regression in Django 1.7
Under Django 1.7, if get_model gets called during app startup and the
models for the requested app just hasn't been loaded yet, our get_model
call fails with a LookupError.
Most of the time, it would be safe to just import the requested models,
and this is what this fix effectively does. This aligns the behaviour of
the old (pre-Django 1.7) and new get_model code again.

This is tracked as issue #1494.