Remove stale symlinked for image_not_found.jpg
I was surprised to find this symlink, as we have a feature to
automatically symlink it: 1c151f6701bc7ff67e963f6501c9afcd4eb43f3d

Furthermore, that functionality was broken by this symlink, as the link
was broken since Oscar was moved to /src. That meant after make sandbox
on a fresh checkout, Oscar would throw the error to ask for proper
symlinking.

The last change to this symlink is much older than the automatic
linking, so I'm hopeful it's safe to delete: 16a5b02aa52