Properly serialise phone number when storing in session
Previously, we just used force_text() to serialise it.
That is the wrong approach, as we allow changing the standard
display to a regional format. But in the session we should always
include the country code. Luckily the easy fix is to just store things
as an international phone number.