Update English .po file
We're asking an editor to review Oscar's documentation and English
translatables. To make that more convenient (i.e. not require having to
get the project set up), it helps to have the English locale files
up-to-date. This will be overridden by the next release anyway, and
shouldn't cause any harm.