Cache Category.get_absolute_url
When we stopped storing the concatenated slugs on each category,
generating the URL for an individual category got more expensive. But
luckily it is safe to naively cache it, as the matching view only
considers the primary key any way. So even a stale cache should lead to
the right category.