Delegate widget initialization only for window.document
initWidgets() is used by JS code here and there for newly added
elements, in that case the elements must be initialized immediately, as
the code sometimes uses those elements right away and relies on them
being initialized.

In particular a newly added datetime widget that wasn't initialized
properly was automatically filled with a date/time, which used the
default locale, because the initialization hadn't happened yet.

The solution used here is to only use the mouseover delegation when
window.document is passed as the node to be initialized, as this happens
during page load. Any *specific* elements given are initialized
immediately, so the calling context can be sure those widgets are ready.