Introduce new voucher datetime fields, deprecate date_created fields
I think almost always datetime fields are more useful than date fields.
We wanted to be able to sort by last created vouchers, which is
not useful when only the day of the year is available.

I'm being a bit handwavy with the migration procedure, because I think
it's overall quite simple to implement, and my guess is that most
deployments don't really care about creation dates for older vouchers.
Hence just doing the switch when Oscar 1.3 comes around should be
sufficient.