Fix errors with shipping discounts on free weightbased shipping.
The Repository.apply_shipping_offer() method returns the shipping method
object when there is no additional discount for an order. This however
broke with the WeightBased, because it's implementation does not work.
self.offer is never assigned, nor useful (a copy-paste error?).

Instead, it should work like the non-model `methods.Base` that also
states the method does not have a default shipping discount.