Enforce order of basket, order, wishlist lines
When working on an Oscar project, I noted that the order of the basket
lines was different from the order of the order lines created from that
basket. This made me investigate how exactly Oscar approaches this.

It turned out that Oscar had no notion of ordering those lines. That
means, the order of them was determined by the database. That will work
in most cases, because at least PostgreSQL tends to return them in the
order of creation - but this is not guaranteed. From
http://www.postgresql.org/docs/9.1/static/queries-order.html:

> If sorting is not chosen, the rows will be returned in an unspecified
> order. The actual order in that case will depend on the scan and join
> plan types and the order on disk, but it must not be relied on. A
> particular output ordering can only be guaranteed if the sort step is
> explicitly chosen.

Luckily, Django's AutoIntegerField guarantees that the primary keys are
incremented. So to get the behaviour we most likely want, we just
enforce ordering by primary keys. I also documented how one would go
about switching to an Amazon-style handling. Implementing that is left
as an exercise to the reader.

As a note: I'm not sure what subtle bugs we previously had when applying
offers in the case of the database returning the lines out-of-order.
Let's not fret on it too much, it's fixed now ;)