Make SimpleAddToBasketForm easier to customize
Upon implementing previous commit, my changes weren't picked up because
Oscar served the SimpleAddToBasketForm. If a user wants to implement
changes in both forms (which is more likely than it is not), the
previous structure meant they had to be duplicated or mixed in. Now, for
most changes, SimpleAddToBasketForm can be ignored.