Remove metadata and notfications context processor
Remove customer email and notifcation apps

This removes the email and notifcation model of the customer app.
The Email model remains from the removing from the cummunication type event
feature. Also this includes a bunch of email templates for this kind of events.

Remove the notifcation model and views because they are not needed neither of us.

Updated po's to the current changes.

Remove the context processor that only pass some variables what are not used.
For the moment I hardcoded the text inside of the layout.html (shop_name)
because I don't wanted to migrate the templates files over also removed the
google analytics templates.