Allow higher version of django-phonenumber-field
This achieves two things: ensure users can pick a Django 3.2 compatible
release, and ensure they can opt to not use version 3.0, which can cause
issues because it adds a ValueError in unexpected places.

Django 3.2 support is scheduled to be ported to Oscar 2.1 LTS. And
phone number field version 3.0 is causing serious issues in existing
codebases, because previous versions allowed incorrect phone numbers to
be stored. But now it prevents saving e.g. addresses again because the
phone number is not considered valid any more.
And the way where the ValueError is raised results in Oscar's checkout
breaking hard with an TransactionManagementError instead of just
displaying an error to enter a valid phone number.